pyproject.toml
[project]
name = "python-genai-practise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.6.2",
    "fastapi>=0.135.1",
    "faster-whisper>=1.2.1",
    "google-genai>=1.66.0",
    "groq>=0.37.1",
    "ipykernel>=7.2.0",
    "kokoro-onnx>=0.5.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-neo4j>=0.7.0",
    "langchain-ollama>=1.0.1",
    "langchain-qdrant>=1.1.0",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.2",
    "langgraph-checkpoint-mongodb>=0.3.1",
    "mem0ai>=1.0.5",
    "ollama>=0.6.1",
    "openai>=2.24.0",
    "pyaudio>=0.2.14",
    "pydantic>=2.12.5",
    "pymongo>=4.15.5",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
    "speechrecognition>=3.15.1",
    "tiktoken>=0.12.0",
    "torch>=2.10.0",
    "transformers>=5.3.0",
    "uvicorn>=0.41.0",
]

[tool.uv.workspace]
members = [
    "pydantic",
]
